% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_loader.R
\name{add_pca_embed}
\alias{add_pca_embed}
\title{Append PCA embeddings and loadings to RDS matrix}
\usage{
add_pca_embed(seurat_object, path_map)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{path_map}{ProcessRun's path map (paths of relevant files)}
}
\value{
Seurat object with PCA embeddings and loadings
}
\description{
Append PCA embeddings and loadings to RDS matrix
}
