% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_loader.R
\name{add_umap_embed}
\alias{add_umap_embed}
\title{Append UMAP embeddings to RDS matrix}
\usage{
add_umap_embed(seurat_object, path_map)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{path_map}{ProcessRun's path map (paths of relevant files)}
}
\value{
Seurat object with UMAP embeddings
}
\description{
Append UMAP embeddings to RDS matrix
}
