% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_prep.R
\name{downsample_counts}
\alias{downsample_counts}
\title{Downsample counts based on downsampled meta}
\usage{
downsample_counts(seurat_obj, downsampled_meta, features = NULL)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{downsampled_meta}{Metadata dataframe, output of downsample_metadata()}

\item{features}{Dataframe of mapping from Ensembl ID to gene name}
}
\value{
Column-wise (barcode-wise) subset of counts matrix
}
\description{
Downsample counts based on downsampled meta
}
