% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_prep.R
\name{downsample_metadata}
\alias{downsample_metadata}
\title{Scale down metadata by a downsample factor}
\usage{
downsample_metadata(
  seurat_obj,
  downsample_factor = 100,
  rounding = "up",
  group_by = "Subclass_Cell_Identity"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{downsample_factor}{By how many times to reduce number of cells per group}

\item{rounding}{Rounding after devision by \code{downsample_factor} ("up", "down" or NULL for standard rounding)}

\item{group_by}{Down-sample per defined group (usually, column name that indicates cell type)}
}
\value{
Row-wise subset of the given metadata dataframe
}
\description{
Scale down metadata by a downsample factor
}
