% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spec.R
\docType{data}
\name{example_spec_py}
\alias{example_spec_py}
\title{Python spec definition}
\format{
An object of class \code{list} of length 2.
}
\usage{
example_spec_py
}
\description{
Converts of Python list of dicts into a reticulate object
}
\keyword{datasets}
