% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spec.R
\docType{data}
\name{example_spec_r}
\alias{example_spec_r}
\title{R spec definition}
\format{
An object of class \code{list} of length 2.
}
\usage{
example_spec_r
}
\description{
List of names lists will convert to list of dictionaries in Python
}
\keyword{datasets}
