% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_loader.R
\name{get_seurat_object}
\alias{get_seurat_object}
\title{Load RDS matrix and embeddings at process run}
\usage{
get_seurat_object(cf_branch)
}
\arguments{
\item{cf_branch}{CellForest branch at selected process}

\item{dim_reduc_funs}{List of dimensionality reduction embeddings to append}
}
\value{
Seurat object with cached dimensionality reduction embeddings
}
\description{
Loads RDS from most recent layer (including current layer).
If RDS comes from parental folder, append embeddings from the current layer.
}
\examples{
library(reticulate)
library(Seurat)

library(cellforestR)
cellforest <- import("cellforest")

root_dir <- "tests/data/example_usage/root"
cf_branch <- cellforest$load(root_dir, branch_spec = example_spec_py)
cf_branch$goto_process("reduce")  # put process or alias (if exists)
seurat_obj <- get_seurat_object(cf_branch)
}
