% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{processes_between_paths}
\alias{processes_between_paths}
\title{Processes that happen between two paths}
\usage{
processes_between_paths(path_upstream, path_downstream)
}
\arguments{
\item{path_upstream}{String path of file upstream (e.g., normalize's RDS)}

\item{path_downstream}{String path of file downstream (e.g., reduce's meta.tsv)}
}
\description{
Processes that happen between two paths
}
