% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_prep.R
\name{save_counts}
\alias{save_counts}
\title{Save prepared counts matrix}
\usage{
save_counts(downsampled_counts, save_dir)
}
\arguments{
\item{downsampled_counts}{Counts matrix}

\item{save_dir}{Folder into which to save counts.tsv}
}
\description{
Save prepared counts matrix
}
