% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_prep.R
\name{save_meta}
\alias{save_meta}
\title{Save prepared metadata}
\usage{
save_meta(downsampled_metadata, save_dir)
}
\arguments{
\item{downsampled_metadata}{metadata dataframe}

\item{save_dir}{Folder into which to save meta.tsv}
}
\description{
Save prepared metadata
}
